Fix integer width mismatch in testsuite

+++ testbench/src/testmain.c
@@ -84,8 +84,9 @@ round_robin_test(unsigned long long seed, unsigned long long skip,
         snprintf(teststring, sizeof teststring,
                  "levelteleport,\"%d\",genesis,\"monsndx #%d\","
                  "wish,\"Z - otyp #%d\",%s,fight,fight,cast,zap,read,drink,"
-                 "fight,fight,wait,wait,%s,wait,wait,wait", (seed + nt) % 50,
-                 mon, item + 1, testable_commands[cmd], testable_commands[cmd]);
+                 "fight,fight,wait,wait,%s,wait,wait,wait",
+                 (int)((seed + nt) % 50), mon, item + 1,
+                 testable_commands[cmd], testable_commands[cmd]);
         (skipping ? skip_test_game : play_test_game)(teststring, verbose);
 
     continue_main_loop:;
